+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Rational
 
+## 0.2.0 (2022-11-24)
+
+- [feat]: `.add()` method for vector addition;
+- [feat]: `.scale()` method for scalar vector multiplication;
+- [feat]: `inverted` property for inverting the vector;
+- [feat]: `.sub()` method for vector subtraction;
+
 ## 0.1.0 (2022-11-23)
 
 - [chore]: First release, just parsing user input functionality;